Harden persistence review follow-ups
Address the verified review findings across the libSQL and PostgreSQL
backends, then fix the smaller regressions that surfaced while replaying
`make all`.

- make libSQL conversation and sandbox decoders fail on corrupt UUIDs
  instead of silently defaulting
- tighten pagination and ordering semantics for conversation listings and
  previews
- move libSQL conversation writes onto transactions where message insert
  and touch must stay atomic
- split oversized history/session modules into dedicated test and helper
  files
- make workspace chunk indices non-negative in the Rust domain model and
  validate database conversions at the boundary
- keep NEAR AI API keys in session state and secrets storage instead of
  mutating process environment
- extend PostgreSQL coverage for routine and LLM call persistence, and
  keep the adapter tests compiling after the cleanup

+++ src/channels/web/handlers/chat_history.rs
@@ -40,7 +40,7 @@ async fn load_stored_history(
     limit: usize,
 ) -> Result<HistoryResponse, (StatusCode, String)> {
     let (messages, has_more) = store
-        .list_conversation_messages_paginated(thread_id, before_cursor, limit as i64)
+        .list_conversation_messages_paginated(thread_id, before_cursor, limit)
         .await
         .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
     let oldest_timestamp = messages
